@@ -12,45 +12,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.4",
+    "@angular/animations": "^14.2.4",
     "@angular/cdk": "^13.2.4",
-    "@angular/common": "^13.2.4",
-    "@angular/compiler": "^13.2.4",
-    "@angular/core": "^13.2.4",
+    "@angular/common": "^14.2.4",
+    "@angular/compiler": "^14.2.4",
+    "@angular/core": "^14.2.4",
     "@angular/fire": "^7.2.1",
-    "@angular/forms": "^13.2.4",
-    "@angular/localize": "^13.2.4",
+    "@angular/forms": "^14.2.4",
+    "@angular/localize": "^14.2.4",
     "@angular/material": "^13.2.4",
     "@angular/material-moment-adapter": "^13.3.3",
-    "@angular/platform-browser": "^13.2.4",
-    "@angular/platform-browser-dynamic": "^13.2.4",
-    "@angular/router": "^13.2.4",
+    "@angular/platform-browser": "^14.2.4",
+    "@angular/platform-browser-dynamic": "^14.2.4",
+    "@angular/router": "^14.2.4",
     "@mapbox/node-pre-gyp": "^1.0.8",
     "@tailwindcss/forms": "^0.4.0",
     "core-js": "^3.20.0",
     "firebase": "^9.6.1",
     "moment": "^2.29.4",
     "rxjs": "~7.4.0",
-    "tslib": "^2.3.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.5",
-    "@angular-eslint/builder": "13.1.0",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/schematics": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "^13.2.5",
-    "@angular/compiler-cli": "^13.2.4",
-    "@angular/language-service": "^13.2.4",
+    "@angular-devkit/build-angular": "^14.2.4",
+    "@angular-eslint/builder": "14.1.2",
+    "@angular-eslint/eslint-plugin": "14.1.2",
+    "@angular-eslint/eslint-plugin-template": "14.1.2",
+    "@angular-eslint/schematics": "14.1.2",
+    "@angular-eslint/template-parser": "14.1.2",
+    "@angular/cli": "^14.2.4",
+    "@angular/compiler-cli": "^14.2.4",
+    "@angular/language-service": "^14.2.4",
     "@types/jasmine": "~3.10.2",
     "@types/jasminewd2": "^2.0.10",
     "@types/node": "~17.0.0",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
-    "autoprefixer": "^10.4.2",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "5.37.0",
+    "@typescript-eslint/parser": "5.37.0",
+    "autoprefixer": "^10.4.12",
+    "eslint": "^8.23.1",
     "eslint-plugin-import": "latest",
     "eslint-plugin-jsdoc": "latest",
     "eslint-plugin-prefer-arrow": "latest",
@@ -61,11 +60,12 @@
     "karma-coverage-istanbul-reporter": "~3.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "postcss": "^8.4.6",
+    "postcss": "^8.4.16",
     "protractor": "~7.0.0",
     "schematics-scss-migrate": "^1.3.14",
     "tailwindcss": "^3.0.22",
     "ts-node": "~7.0.0",
-    "typescript": "^4.5.4"
+    "tslib": "^2.4.0",
+    "typescript": "^4.8.4"
   }
 }
